requirements.txt
tweepy==3.9.0
google-cloud-storage==1.30.0
nltk==3.6.6
Google-Images-Search==1.3.10